+++ src/AngleSharp/Dom/DocumentReadyState.cs
@@ -1,5 +1,7 @@
-﻿namespace AngleSharp.Dom
+namespace AngleSharp.Dom
 {
+    using AngleSharp.Attributes;
+
     /// <summary>
     /// Defines the document readiness.
     /// </summary>
@@ -8,14 +10,17 @@
         /// <summary>
         /// The document is still loading.
         /// </summary>
+        [DomName("loading")]
         Loading,
         /// <summary>
         /// The document is interactive, i.e. interaction possible.
         /// </summary>
+        [DomName("interactive")]
         Interactive,
         /// <summary>
         /// Loading is complete.
         /// </summary>
+        [DomName("complete")]
         Complete
     }
 }